% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{altSourceToStimulant}
\alias{altSourceToStimulant}
\title{A function to convert altSource inputs to stimulants.}
\usage{
altSourceToStimulant(hormones)
}
\arguments{
\item{hormones}{the list of hormones contained in a network object.}
}
\description{
A function to convert altSource inputs to stimulants.
}
