% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{buildEquation}
\alias{buildEquation}
\title{A function to generate equations from a network object}
\usage{
buildEquation(network, upPenalty, containerPenalty, file = "./equations.R")
}
\arguments{
\item{network}{an object of class network}

\item{upPenalty}{a penalty applied to hormones traveling upwards. Behaves as
a multiplier.}

\item{containerPenalty}{a penalty applied to hormones traveling upwards.
Behaves as a multiplier.}

\item{file}{the name of the file that you want the equations to be saved to.
The default is to create an R script called equations.R in the
current working directory.}
}
\description{
A function to generate equations from a network object
}
