% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNintegration.R
\name{buildHormone}
\alias{buildHormone}
\title{A function to build a hormone object}
\usage{
buildHormone(nodeInfo, arcInfo, i, logicIndex, ids, nodesList, lang)
}
\arguments{
\item{nodeInfo}{a data.frame containing summary information for nodes}

\item{arcInfo}{a data.frame containing arc information, including
origins and destinations}

\item{i}{external index}

\item{logicIndex}{a vector consisting of the indexes of logical
operators}

\item{ids}{a vector of all the ids}

\item{nodesList}{a list of all the nodes in the diagram}

\item{lang}{the SBGN language used to construct the original diagram}
}
\description{
Organises information into the correct format to describe a hormone
}
