% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{buildModel}
\alias{buildModel}
\title{A function to generate equations from a network object}
\usage{
buildModel(network, file = "./equations.R")
}
\arguments{
\item{network}{an object of class network}

\item{file}{the name of the file that you want the equations to be saved to.
The default is to create an R script called equations.R in the
current working directory.}
}
\description{
This function is used to define the script that is needed to simulate the
model of interest. It defines the genotypes of the network, and sets them
to a default of 1. It also initiates that data.frame which will record the
changes to the system over time, including initiating the starting values
for the nodes. The default starting values are 1. A nextStep function is
also defined, which provides the information needed to calculate the next
time step in the simulation. This file can be edited by the user as required.
The file will need to be provided to the simulateNetwork function as an input.
}
