% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{buildNetwork}
\alias{buildNetwork}
\title{A function for building a network object}
\usage{
buildNetwork(hormones, genotypes, name)
}
\arguments{
\item{hormones}{a list containing objects of class hormone.}

\item{genotypes}{a list containing objects of class hormone.}

\item{name}{the name to be given to the network}
}
\description{
This function accepts a list of nodes and a list of genotypes. It will check
if the objects are appropriate, and organise them into a single object of
class network. Will check the network to make sure that all objects and
genotypes have been pointed to, and will return a list of objects that have
not been.
}
\examples{
NA
}
