% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustnessChecks.R
\name{checkEdgeTypes}
\alias{checkEdgeTypes}
\title{A function to test how network behavior changes in response to changing the
nature of edge types.}
\usage{
checkEdgeTypes(startingNetwork)
}
\arguments{
\item{startingNetwork}{an object of class network describing}
}
\description{
A function to test how network behavior changes in response to changing the
nature of edge types.
}
