% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluation.R
\name{consistencyCheck}
\alias{consistencyCheck}
\title{A function to check the internal consistency of a Network object.}
\usage{
consistencyCheck(network)
}
\arguments{
\item{network}{an object of class network.}
}
\description{
This function checks the inputs and outputs of all nodes in the
network to make sure that all inputs and outputs for nodes are
in agreement with each other. In addition, checks that the
genotypes are consistent in their coregulators, and that they
agree with the nodes in terms of regulation.
}
