% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{convertSBGNdiagram}
\alias{convertSBGNdiagram}
\title{A function that converts an SBGN-ML text file into a network object}
\usage{
convertSBGNdiagram(file, path = ".", networkName)
}
\arguments{
\item{file}{A string. A text file with extension .xml containing the markdown output
of an sbgn diagram.}

\item{path}{A string. The path directory for where to store the network.}

\item{networkName}{A string. The name that you wish to give the network.}
}
\description{
A function that converts an SBGN-ML text file into a network object
}
