% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{coregulators}
\alias{coregulators}
\title{A function that creates a string with coregulators multiplied together}
\usage{
coregulators(coreg, returnNum = FALSE)
}
\arguments{
\item{returnNum}{return the number of unique coregulator sets. Default is
set to FLASE.}

\item{coregulators}{a subsed of a the inputs of a hormone object. This subset
must all have the same modulating effect, and have coregulators}
}
\description{
Coregulators are sorted so that their multiplications are only represented
once. Each modulator is given a temporal modifier. Sets of coregulators are
summed.
}
