% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{exogenousScreen}
\alias{exogenousScreen}
\title{A function to generate all combinations of two vectors containing the values
to be screened for two nodes. This function will also include the possibility
of no exogenous values being provided in the form of NA.}
\usage{
exogenousScreen(nodes, screen1, screen2)
}
\arguments{
\item{nodes}{a vector containing the names of two nodes of interest}

\item{screen1}{a vector containing all the values to be tested for the
first listed node.}

\item{screen2}{a vector containing all the values to be tested for the
second listed node.}
}
\description{
A function to generate all combinations of two vectors containing the values
to be screened for two nodes. This function will also include the possibility
of no exogenous values being provided in the form of NA.
}
