% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustnessChecks.R
\name{exploreEdges}
\alias{exploreEdges}
\title{A function to test how network behavior changes in response to changing the
nature of edge types.}
\usage{
exploreEdges(
  startingNetwork,
  folder,
  tmax = 100,
  steadyThreshold = 4,
  ruleStyle = "Dun",
  nesStimStyle = "Linear",
  nesStimFile = NULL,
  saveNetwork = T,
  saveOutput = F
)
}
\arguments{
\item{startingNetwork}{an object of class network describing the system to
to be tested}

\item{folder}{the directory containing the original nextStep function built
on the original network}

\item{tmax}{description}

\item{steadyThreshold}{description}

\item{ruleStyle}{description}

\item{nesStimStyle}{description}

\item{nesStimFile}{description}

\item{saveNetwork}{description}

\item{saveOutput}{description}
}
\description{
A function to test how network behavior changes in response to changing the
nature of edge types.
}
