% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{finalStates}
\alias{finalStates}
\title{A function to pull the final states from simulations. This function will work
for output from bothe the simulateNetwork, and setupSims functions.}
\usage{
finalStates(simulations)
}
\arguments{
\item{simulations}{a list containing the output of the setupSims function}
}
\description{
A function to pull the final states from simulations. This function will work
for output from bothe the simulateNetwork, and setupSims functions.
}
