% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{generateAjacencyMatrix}
\alias{generateAjacencyMatrix}
\title{A function to generate an adjacency matrix from a network object}
\usage{
generateAjacencyMatrix(network, addGenes = F, nGenes = NA)
}
\arguments{
\item{network}{an onject of class network generated from the
convertSBGNdiagram function.}

\item{addGenes}{logical defaulting to FALSE. States if additional
incoming nodes should be added to and node in the
network that has a modifier.}

\item{nGenes}{in the case that additional nodes will be added,
how many should be added.}
}
\description{
A function to generate an adjacency matrix from a network object
}
