% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{generateC}
\alias{generateC}
\title{A function to generate a C script which will execute a simulation of the
network given a starting condition.}
\usage{
generateC(
  network,
  tmax = 100,
  steadyThreshold = 4,
  folder = "./Model",
  forceOverwrite = FALSE
)
}
\arguments{
\item{network}{an object of class network.}

\item{tmax}{the maximum value that the simulation will be allowed to
proceed. If the midpoint is reached, a warning will be returned.
The default value is set to 0.}

\item{steadyThreshold}{the number of decimal places to which node values must
be equivalent to be considered a steady state. This threshold must
be passed for all nodes.}

\item{folder}{the name of the folder that you want the components of the
model to be saved to. The default is to create a directory
called Model in the current working directory. The user can
provide their own directory and folder name. In this folder
will be generated three R scripts: one to define genotypes
(genotypeDef.R), one to define the starting node values
(nodestartDef.R), and one to define the function that
gives the difference equations that are used to simulate the
network (nextStep.R).}

\item{forceOverwrite}{default set to FALSE. Will stop the function if the
folder already exits. Can set to true if you want to replace
the existing folder.}
}
\description{
A function to generate a C script which will execute a simulation of the
network given a starting condition.
}
