% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generateColours}
\alias{generateColours}
\title{A function that generates colours for a vector with numeric values.}
\usage{
generateColours(x, maxvalue = NA)
}
\arguments{
\item{x}{a vector of numerical values containing node expression levels as
normalised by the wild type (1 is the baseline expression).}

\item{maxvalue}{the value that the user wants to be treated as the maximum
over}
}
\description{
Generates hex colours for relative expression data. WT (1) values are white,
complete lack of expression is black. Overexpressed nodes will be given a
gradient from white to red. Underexpressed nodes will be given a gradient from white to blue.
}
