% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{generateEdgeList}
\alias{generateEdgeList}
\title{A function to generate a data frame which lists all of
the edges of a network.}
\usage{
generateEdgeList(
  network,
  keepAltSource = F,
  keepNecessaryStimulant = F,
  recordConjunctions = F
)
}
\arguments{
\item{network}{an object of class Network.}

\item{keepAltSource}{default set to FALSE. Specify if
you want to recognise if the incoming node
is an alternate source of the same 'hormone' type.
If not, these alternative sources will be reported
as stimulants.}

\item{keepNecessaryStimulant}{default set to FALSE.
Specify if you want to distinguish necessary
stimulants from stimulants.}

\item{recordConjunctions}{default set to FALSE.
Specify if you want to record any conjunctions
that appear in the network.}
}
\description{
A function to generate a data frame which lists all of
the edges of a network.
}
