% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{generateEquation}
\alias{generateEquation}
\title{A function to generate an equation for a specific node}
\usage{
generateEquation(node, genotypes, language, style = "Dun")
}
\arguments{
\item{node}{a node from within a network object}

\item{language}{which programming language should the equation be generated in?
Can be either "R", or "C".}

\item{style}{either "Dun", or "Mike"}

\item{genes}{the list of genes frome within a network object}
}
\description{
A function to generate an equation for a specific node
}
