% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{generateGeneToNodeList}
\alias{generateGeneToNodeList}
\title{a function to map the genes that effect nodes in a network}
\usage{
generateGeneToNodeList(network)
}
\arguments{
\item{network}{and object of class network}
}
\description{
a function to map the genes that effect nodes in a network
}
