% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\docType{class}
\name{Genotype-class}
\alias{Genotype-class}
\alias{Genotype}
\title{A function to generate objects of class 'Hormone'}
\description{
A function to generate objects of class 'Hormone'
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the gene. Will be used as a tag to place influence within the network.}

\item{\code{expression}}{data.frame with column names Compartment, and Expression.
The compartment column can include "scion", or "rootstock".
The expression column is a numerical value set between 0 (no expression),
and 1 (full expression). All expression levels are set to the default 1
(the wildtype genotype).}

\item{\code{coregulator}}{of class character. The names of other genotypes that coregulate together.}

\item{\code{influence}}{a data.frame with the column names Node and Influence. Influences can one
of either "production", "degradation", "inhibition", "perception".}
}}

