% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{genotypeScreen}
\alias{genotypeScreen}
\title{A function to generate genotype screenings.}
\usage{
genotypeScreen(folder, maxMutations = 1)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{maxMutations}{the maximum number of concurrent mutations in a given
compartment.}
}
\description{
Will generate genetype screens which considers compartmental information,
allows for the possibility of multiple mutations.
}
