% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNintegration.R
\name{getSubmapNodes}
\alias{getSubmapNodes}
\title{A function to pull out information from with in submaps}
\usage{
getSubmapNodes(node, comp, newNodes = NA)
}
\arguments{
\item{node}{the parent node containing a submap}

\item{comp}{the compartment within which the submap sits}

\item{newNodes}{any carryover nodes in the case that the
the function is being used recursively (if there
are submaps within submaps)}
}
\description{
This function pulls out the nodes existing within a submap,
as well as collects the node information for processing.
}
