% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\docType{class}
\name{Hormone-class}
\alias{Hormone-class}
\alias{Hormone}
\title{A function to generate objects of class 'Hormone'}
\description{
A function to generate objects of class 'Hormone'
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the hormone. Will be used as a tag to place node within the network.}

\item{\code{container}}{the location of action for the hormone. Either "scion", or "rootstock".}

\item{\code{inputs}}{data.frame with column names Node, and Influence. Influences can one of either
"stimulation", "inhibition", "necessary stimulation", "necessary inhibition",
"sufficient stimulation", "sufficient inhibition", "Unknown".}

\item{\code{outputs}}{data.frame with column names Node, and Influence. Influences can one of either
"stimulation", "inhibition", "necessary stimulation", "necessary inhibition",
"sufficient stimulation", "sufficient inhibition", "Unknown".}

\item{\code{travel}}{specifies if the hormone travels between compartments. Hormones travel from
their current container, to the other. Is a numeric value specifying the rate of
travel. Is 0 if there is no travel between compartments. Default for hormones
traveling down is 1, and up is 0.8 to reflect that signals moving up take a
longer time than signals moving down.}

\item{\code{degradation}}{the rate at which the hormone degrades.}

\item{\code{genotypes}}{a vector with objects of class character. Lists the genotypes that
are important to this node.}
}}

\examples{
NA
}
