% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inhibitor}
\alias{Inhibitor}
\title{Definition for the Inhibitor object.}
\format{
A list containing 6 objects:
\describe{
  \item{name}{name of object}
  \item{container}{location of action}
  \item{inputs}{influencing nodes}
  \item{outputs}{influenced nodes}
  \item{travel}{speed of travel}
  \item{degradation}{rate of degradation of product}
  \item{genotypes}{controlling genotypes}
  ...
}
}
\usage{
Inhibitor
}
\description{
An object of class hormone.
}
\keyword{datasets}
