% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{langConversion}
\alias{langConversion}
\title{A data.frame containing the language conventions and their
equivalence of the peaSoup package, and the SBGN languages
'activity flow' and 'entity relationship'.}
\format{
a data.frame containing three rows:
\describe{
  \item{peaSoup}{The relationships supported by peaSoup}
  \item{AF}{The equivalent relationship for the activity flow notation}
  \item{ER}{The equivalent relationship for the entitiy relationship notation}
}
}
\usage{
langConversion
}
\description{
An object of class data.frame
}
\keyword{datasets}
