% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{languageConversion}
\alias{languageConversion}
\title{A function to convert SBGN language to peaSoup language}
\usage{
languageConversion(influenceFrame, language)
}
\arguments{
\item{influenceFrame}{the column of either the influence slot of a hormone object}

\item{language}{the SBGN language style used to build an initial diagrammatic network}
}
\description{
A function to convert SBGN language to peaSoup language
}
