% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{listGenotypes}
\alias{listGenotypes}
\title{A function to list all objects of class genotype in the current environment}
\usage{
listGenotypes()
}
\description{
A function to list all objects of class genotype in the current environment
}
