% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{loadMyModel}
\alias{loadMyModel}
\title{A function to restore a constructed model.}
\usage{
loadMyModel(wd = ".")
}
\description{
All hormone and genotype objects will be purged from the environment,
and the objects of the desired model will be loaded.
}
