% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\alias{Network}
\title{A function to generate objects of class 'Network'}
\description{
A function to generate objects of class 'Network'
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the network.}

\item{\code{model}}{a list of length 2 with the names Hormones and Genotypes}
}}

