% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{numCombn}
\alias{numCombn}
\title{A function to calculate the number of unique combinations of a vector
of length n of sample size r.}
\usage{
numCombn(n, r)
}
\arguments{
\item{n}{number of items in the vector.}

\item{r}{the size of the sample.}
}
\description{
A function to calculate the number of unique combinations of a vector
of length n of sample size r.
}
