% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaNetwork}
\alias{peaNetwork}
\title{Definition for the peaNetwork object.}
\format{
A list containing 4 objects:
\describe{
  \item{name}{name of object}
  \item{expression}{location of action}
  \item{coregulator}{coregulating genotypes}
  \item{influence}{influenced nodes}
}
}
\usage{
peaNetwork
}
\description{
An object of class Network
}
\keyword{datasets}
