% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.simData}
\alias{plot.simData}
\title{A function to comparatively plot the outcomes of different simulations}
\usage{
\method{plot}{simData}(data, node, bioData = NA, ...)
}
\arguments{
\item{data}{a data.frame containing the outcome of a set of simulations. The
first row should contain the wild type condition.}

\item{node}{a string showing which node data will be plotted.}

\item{bioData}{a matrix containing the biological data corresponding to the
simulated data.}

\item{cols}{specifies the colours that will be used to plot node values
over time. Can accept colour palettes from the viridis package.
Colours can be individually selected and provided as a vector
of the same length as the nodes that you wish to plot. If no
value is provided will use the viridis palette as default.}
}
\description{
This function allows you to quickly plot the outcome of different simulations.
}
