% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\docType{data}
\name{print.Network}
\alias{print.Network}
\title{This is a function which specifies the print method for objects
of class Network.}
\format{
An object of class \code{character} of length 1.
}
\usage{
print.Network
}
\description{
This method is applied automatically with out needing to call
the print function.
}
\examples{
NA
}
\keyword{datasets}
