% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{quickPlot}
\alias{quickPlot}
\title{A function to quickly inspect the through time outcome of a simulation}
\usage{
quickPlot(simulationData)
}
\arguments{
\item{simulationData}{the data.frame output of a simulation.}
}
\description{
A function to quickly inspect the through time outcome of a simulation
}
