% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{randomStartScreen}
\alias{randomStartScreen}
\title{A function to generate a number of random starting point reassignments.}
\usage{
randomStartScreen(folder, restarts, minVal = 0, maxVal = 2, static = FALSE)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{restarts}{the number of randomly assigned starting points that you
want to test.}

\item{minVal}{default set to 0. The minimum value for a node to be.}

\item{maxVal}{default set to 2. The maximum value for a node to be.}

\item{static}{default set to FALSE. Determines if the final data.frame is
saved as a .Rdata file. The default condition overwrites the
nodestartDefinition.R file to define the new dat data.frame.}
}
\description{
Starting points will be generated for each node, pulled from a uniform
distribution. The existing starting values in the model folder will not
be overwritten.
}
