% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{report}
\alias{report}
\title{A function to produce a report of simulation progression}
\usage{
report(x, r)
}
\arguments{
\item{x}{the index for the simulation completed}

\item{r}{the total number of simulations to be completed}
}
\description{
A function to produce a report of simulation progression
}
