% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{restoreBaseModel}
\alias{restoreBaseModel}
\title{A function to restore the base model.}
\usage{
restoreBaseModel()
}
\description{
All hormone and genotype objects will be purged from the environment,
and the standard objects reloaded.
}
