% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RMS4}
\alias{RMS4}
\title{Definition for the RMS4 object.}
\format{
A list containing 4 objects:
\describe{
  \item{name}{name of object}
  \item{expression}{location of action}
  \item{coregulator}{coregulating genotypes}
  \item{influence}{influenced nodes}
}
}
\usage{
RMS4
}
\description{
An object of class genotype.
}
\keyword{datasets}
