% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjManipulate.R
\name{saveNetwork}
\alias{saveNetwork}
\title{A function to save network objects created by the user}
\usage{
saveNetwork(network, file, overwrite = FALSE, readOnly = TRUE)
}
\arguments{
\item{network}{an object of class network.}

\item{overwrite}{logical. Specifies if you want to be able to overwrite
files if they have the same name. Default set to false.}

\item{readOnly}{logical. Specifies if you would like to set the file
permissions on the saved object to read only. Default
set to true.}

\item{path}{a string. The path directory for where to store the network.}
}
\description{
Saves network objects to a dedicated 'Networks' file, and tries to prevent
loss of information.
}
