% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{setupSims}
\alias{setupSims}
\title{A wrapper function that allows you to run simulations of a network under
many different conditions.}
\usage{
setupSims(
  folder,
  delay = 2,
  maxStep = 100,
  steadyThreshold = 4,
  exogenousSupply = FALSE,
  priorScreen = FALSE,
  saveOutput = FALSE,
  robustnessTest = FALSE,
  genotypeBaseline = FALSE,
  nodestartBaseline = FALSE,
  altTopologyName = NULL
)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{delay}{the amount of delay for any delayed transport. Default set to 2.}

\item{maxStep}{the maximum number of timesteps for which to simulate.}

\item{steadyThreshold}{the number of decimal places to which node values must
be equivalent to be considered a steady state. This threshold must
be passed for all nodes.}

\item{exogenousSupply}{specifies if the value of a node (or nodes) is
determined by an outside supply. In this case, the value of the
node is supplied by the user and remains consistent throughout
the course of the simulation. The default value for this
argument is FALSE. If set to TRUE, there must exist an exogenousDef
object in the model folder. This can be generated with the
exogenousScreen function.}

\item{priorScreen}{logical. Specifies if the function should collect
modifier values generated from generated prior distributions.
Default is set to FALSE.}

\item{saveOutput}{logical. Default set to FALSE. Indicates if the output of
simulation screen should be automatically saved in the
provided folder location upon completion.}

\item{robustnessTest}{logical. Defaults to FALSE. Specifies if the nextStep
function being used is part of a network robustness check.}

\item{genotypeBaseline}{logical. Defaults to FALSE. Specified if only the
first row of the genotype object should be used. If selected,
the output object name will reflect that decision. If
genotypeBaseline = TRUE and priorScreen = TRUE, a warning will
be thrown to the user and the priorScreen argument will take
precedence.}

\item{nodestartBaseline}{logical. Defaults to FALSE. Specified if only the
first row of the nodestart object should be used. If selected,
the output object name will reflect that decision.}

\item{altTopologyName}{default to NULL. If robustnessTest = TRUE, this argument
allows the user to keep the generated alternate nextStep function
with a specific name. If no name is provided, the alternate
nextStep functions will be names nextStepAlt.R.}
}
\description{
This function allows you to set up a variety of genotype and starting node
conditions to be simulated. It also allows you to test the outcome of
simulations under randomly assigned starting values. The function will make
sure to include one wild type condition (all genotypes set to 1) regardless
of if specified by the user.
}
