% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluation.R
\name{stabilityVector}
\alias{stabilityVector}
\title{A function to create a vector that reports whether simulations
have managed to reach stability by the maximum timestep.}
\usage{
stabilityVector(allSim)
}
\arguments{
\item{allSim}{the output list of setupSims}
}
\description{
A function to create a vector that reports whether simulations
have managed to reach stability by the maximum timestep.
}
