% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrigolactoneS}
\alias{StrigolactoneS}
\title{Definition for the StrigolactoneS object.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{name}{name of object}
  \item{container}{location of action}
  \item{inputs}{influencing nodes}
  \item{outputs}{influenced nodes}
  \item{travel}{speed of travel}
  \item{genotypes}{controlling genotypes}
  ...
}
}
\usage{
StrigolactoneS
}
\description{
An object of class hormone.
}
\keyword{datasets}
