% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{ANDfuncString}
\alias{ANDfuncString}
\title{A function for generating the code for a conjunction}
\usage{
ANDfuncString(coregString, language, style)
}
\arguments{
\item{coregString}{a string listing coregulators}

\item{language}{programming language. "R" or "C"}

\item{style}{form of mathematical representation of AND}
}
\description{
A function for generating the code for a conjunction
}
