% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{buildModel}
\alias{buildModel}
\title{A function to generate equations from a network object}
\usage{
buildModel(network, folder = "./Model", forceOverwrite = FALSE)
}
\arguments{
\item{network}{an object of class network}

\item{folder}{the name of the folder that you want the components of the
model to be saved to. The default is to create a directory
called Model in the current working directory. The user can
provide their own directory and folder name. In this folder
will be generated three R scripts: one to define genotypes
(genotypeDefinition.R), one to define the starting node values
(nodestartDefinition.R), and one to define the function that
gives the difference equations that are used to simulate the
network (nextStep.R).}

\item{forceOverwrite}{default set to FALSE. Will stop the function if the
folder already exits. Can set to true if you want to replace
the existing folder}
}
\description{
This function is used to define the script that is needed to simulate the
model of interest. It defines the genotypes of the network, and sets them
to a default of 1. It also initiates that data.frame which will record the
changes to the system over time, including initiating the starting values
for the nodes. The default starting values are 1. A nextStep function is
also defined, which provides the information needed to calculate the next
time step in the simulation. This file can be edited by the user as required.
The file will need to be provided to the simulateNetwork function as an input.
}
