% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNintegration.R
\name{convertSBGNdiagram}
\alias{convertSBGNdiagram}
\title{A function that converts an SBGN-ML text file into a network object}
\usage{
convertSBGNdiagram(file, networkName)
}
\arguments{
\item{file}{a string. A text file with extension .xml containing the markdown output
of an sbgn diagram.}

\item{networkName}{a string. The name that you wish to give the network.}
}
\description{
A function that converts an SBGN-ML text file into a network object
}
