% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{differenceString}
\alias{differenceString}
\title{A function to build the difference equation including delays}
\usage{
differenceString(string, delays = NA, takeProduct = FALSE, language)
}
\arguments{
\item{string}{whatever string is being constructed.}

\item{delays}{a vector specifying if there are any delays associated with
a particular input.}

\item{takeProduct}{logical. Should be set to TRUE if this function is being
used to collapse necessary stimulants (* instead of +). Can be set to
NULL of you do not want to concatenate into a single string.}

\item{language}{which programming language should the equation be generated in?
Can be either "R", or "C".}
}
\description{
A function to build the difference equation including delays
}
