% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{finalStates}
\alias{finalStates}
\title{A function to pull the final states from a set of simulations and normalise
against the wildtype (baseline).}
\usage{
finalStates(simulations, nodes = NA)
}
\arguments{
\item{simulations}{a list containing the output of the setupSims function}

\item{nodes}{a vector containing all the nodes that the user wants to plot}
}
\description{
A function to pull the final states from a set of simulations and normalise
against the wildtype (baseline).
}
