% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{genotypeScreen}
\alias{genotypeScreen}
\title{A function to generate genotype screenings.}
\usage{
genotypeScreen(
  folder,
  maxMutations = 1,
  mutationVals = 0,
  returnExcel = FALSE,
  static = TRUE
)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{maxMutations}{the maximum number of concurrent mutations in a given
compartment.}

\item{mutationVals}{for each combination of mutated genes, what value should
the genes take on? Default is set to 0 (knockout). Can provide a
vector for the variety of values that are wanted. Only one mutation
value will be considered at a time.}

\item{returnExcel}{default set to FALSE. If TRUE, will create an excel file
and save it in the model folder. This file will contain a matrix
showing all of the genotype combinations that have been screened. This
can be used to record any existing experimental outcomes, to be
compared against simulated outcomes later on.}

\item{static}{default set to TRUE. Determines if the final data.frame is
saved as a .Rdata file. The default condition overwrites the
genotypeDef.R file to define the new genotypeDef data.frame.}
}
\description{
Will generate genotype screens which considers compartmental information,
allows for the possibility of multiple mutations.
}
