% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{genotypeScreen}
\alias{genotypeScreen}
\title{A function to generate genotype screens.}
\usage{
genotypeScreen(
  folder,
  numMutations = 1,
  mutationVal = 0,
  returnExcel = FALSE,
  graft = FALSE
)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{numMutations}{how many concurrent mutations do you want to test.}

\item{mutationVal}{for each combination of mutated genes, what value should
the genes take on? Default is set to 0 (knockout).}

\item{returnExcel}{default set to FALSE. If TRUE, will create an excel file
and save it in the model folder. This file will contain a matrix
showing all of the genotype combinations that have been screened. This
can be used to record any existing experimental outcomes, to be
compared against simulated outcomes later on.}

\item{graft}{default set to FALSE. Indicates if the user wants to
consider differential expression between compartments.}
}
\description{
Will generate genotype screens which considers compartmental information,
and allows for the possibility of multiple types of mutation.
}
