% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNintegration.R
\name{languageConversion}
\alias{languageConversion}
\title{A function to convert SBGN language to peaSoup language}
\usage{
languageConversion(influenceFrame, language)
}
\arguments{
\item{influenceFrame}{the influence column of either the input or output
slot of a hormone object}

\item{language}{the SBGN language style used to build an initial diagrammatic
network}
}
\description{
A function to convert SBGN language to peaSoup language
}
