% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{modifierPriorScreen}
\alias{modifierPriorScreen}
\title{A function to generate a screen of modifier conditions based on a set of
prior distributions.}
\usage{
modifierPriorScreen(
  folder,
  priorDistribution = "rlnorm",
  n,
  savePriors = T,
  returnVals = F
)
}
\arguments{
\item{folder}{a string containing the directory of the folder containing
your generated model.}

\item{priorDistribution}{states the prior distribution to be used to generate
modifier values. If of length 1, the prior will be applied to all
modifier values. If length is greater than 1, the vector must be
named with the corresponding node name.}

\item{n}{the number of observations.}

\item{savePriors}{logical. If the prior screen should be saved in the provided
folder.}

\item{returnVals}{logical. If the output should be returned to the user.}
}
\description{
A function to generate a screen of modifier conditions based on a set of
prior distributions.
}
