% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.simData}
\alias{plot.simData}
\title{A function to comparatively plot the outcomes of different simulations}
\usage{
\method{plot}{simData}(data, node, bioData = NA, ...)
}
\arguments{
\item{data}{a data.frame containing the outcome of a set of simulations. The
first row should contain the wild type condition.}

\item{node}{a string showing which node data will be plotted.}

\item{bioData}{a matrix containing the biological data corresponding to the
simulated data.}
}
\description{
This function allows you to quickly plot the outcome of different simulations.
}
