% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{quickPlot}
\alias{quickPlot}
\title{A function to quickly inspect the through time outcome of a simulation}
\usage{
quickPlot(
  simulationData,
  col = NA,
  logTransform = FALSE,
  addLabels = FALSE,
  title = NA
)
}
\arguments{
\item{simulationData}{the data.frame output of a simulation. A subset of the
data.frame can be passed if the user does not want to plot all of the
nodes.}

\item{col}{default is set to NA, in which case the viridis colour palette will
be used. Otherwise, the user can pass a vector of colors the same length
as the number of nodes to be plotted}

\item{logTransform}{whether the simulation values should be log transformed.}

\item{addLabels}{if you want to add labels next to the
final value of each node.}
}
\description{
A function to quickly inspect the through time outcome of a simulation
}
