% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluation.R
\name{reportCondition}
\alias{reportCondition}
\title{A function to report any deviation from the baseline for simulations}
\usage{
reportCondition(sims, returnName = FALSE)
}
\arguments{
\item{sims}{either the full output list returned by setupSims, in which case
the deviations of the full simulations will be returned.
Otherwise, if only a single simulation is provided (in the form
output$screen[[index]]), the deviation of that specific simulation
condition will be returned.}

\item{returnName}{logical. Default set to FALSE. Returns a vector of string
names for the conditions if set to TRUE.}
}
\description{
A function to report any deviation from the baseline for simulations
}
