% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{setupSims}
\alias{setupSims}
\title{A wrapper function that allows you to run simulations of a network under
many differet conditions.}
\usage{
setupSims(file, randomStart = NA, delay = 2, tmax = NA)
}
\arguments{
\item{file}{a file of the type produced by the buildModel function. Contains
the starting values for nodes, defines the genotypes of the
model, and defines the nextStep function for simulating the
change at each timestep.}

\item{randomStart}{the number of random starting point reassignments that
you want to test. Starting points will be generated for each node,
pulled from a uniform distribution ranging from 0 to 2. The default
is set to NA, in which case only the provided start values (as stated
in the provided file) will be run.}

\item{delay}{the amount of delay for any delayed transport. Default set to 2.}

\item{tmax}{the maximum number of timesteps for which to simulate.}
}
\description{
This function allows you to set up a variety of genotype conditions to be
simulated. It also allows you to test the outcome of simulations under
randomly assigned starting values. The function will make sure to include
one wild type condition (all genotypes set to 1) regardless of if specified
by the user.
}
