% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{setupSims}
\alias{setupSims}
\title{A wrapper function that allows you to run simulations of a network under
many differet conditions.}
\usage{
setupSims(folder, delay = 2, tmax = NA)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{delay}{the amount of delay for any delayed transport. Default set to 2.}

\item{tmax}{the maximum number of timesteps for which to simulate.}
}
\description{
This function allows you to set up a variety of genotype conditions to be
simulated. It also allows you to test the outcome of simulations under
randomly assigned starting values. The function will make sure to include
one wild type condition (all genotypes set to 1) regardless of if specified
by the user.
}
