% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{setupSims}
\alias{setupSims}
\title{A wrapper function that allows you to run simulations of a network under
many different conditions.}
\usage{
setupSims(folder, delay = 2, tmax = NA, exogenousSupply = NULL)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{delay}{the amount of delay for any delayed transport. Default set to 2.}

\item{tmax}{the maximum number of timesteps for which to simulate.}

\item{exogenousSupply}{specifies if the value of a node (or nodes) is
determined by an outside supply. In this case, the value of the
node is supplied by the user and remains consistent throughout
the course of the simulation. The default value for this
argument is NULL. To specify nodes with an exogenous supply,
provide a data frame with column names corresponding to the
values of the nodes, with each vector member named after their
respective node.}
}
\description{
This function allows you to set up a variety of genotype and starting node
conditions to be simulated. It also allows you to test the outcome of
simulations under randomly assigned starting values. The function will make
sure to include one wild type condition (all genotypes set to 1) regardless
of if specified by the user.
}
