% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{setupSims}
\alias{setupSims}
\title{A wrapper function that allows you to run simulations of a network under
many different conditions.}
\usage{
setupSims(
  folder,
  delay = 2,
  tmax = NA,
  steadyThreshold = 4,
  exogenousSupply = NULL,
  priorScreen = T,
  saveOutput = F,
  robustnessTest = F,
  altTopologyName = NULL,
  report = F
)
}
\arguments{
\item{folder}{the model folder generated by the buildModel function.}

\item{delay}{the amount of delay for any delayed transport. Default set to 2.}

\item{tmax}{the maximum number of timesteps for which to simulate.}

\item{steadyThreshold}{the number of decimal places to which node values must
be equivalent to be considered a steady state. This threshold must
be passed for all nodes.}

\item{exogenousSupply}{specifies if the value of a node (or nodes) is
determined by an outside supply. In this case, the value of the
node is supplied by the user and remains consistent throughout
the course of the simulation. The default value for this
argument is NULL. To specify nodes with an exogenous supply,
provide a data frame with column names corresponding to the
values of the nodes, with each vector member named after their
respective node.}

\item{priorScreen}{logical. Specifies if the function should collect
modifier values generated from generated prior distributions.}

\item{saveOutput}{logical. Default set to FALSE. Indicates if the output of
simulation screen should be automatically saved in the
provided folder location upon completion.}

\item{robustnessTest}{logical. Defaults to FALSE. Specifies if the nextStep
function being used is part of a network robustness check.}

\item{altTopologyName}{default to NULL. If robustnessTest = T, this argument
allows the user to keep the generated alternate nextStep function
with a specific name. If no name is provided, the alternate
nextStep functions will be names nextStepAlt.R.}

\item{report}{Logical, defaults to FALSE. Indicates if the index of the
completed simulations is returned.}
}
\description{
This function allows you to set up a variety of genotype and starting node
conditions to be simulated. It also allows you to test the outcome of
simulations under randomly assigned starting values. The function will make
sure to include one wild type condition (all genotypes set to 1) regardless
of if specified by the user.
}
