% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelGeneration.R
\name{simulateNetwork}
\alias{simulateNetwork}
\title{A function that simulates the outcomes of network peturbations}
\usage{
simulateNetwork(file, network, nsteps = NA, method = NA)
}
\arguments{
\item{file}{the same file pathway that was used to generate the model with
the buildModel function.}

\item{nsteps}{the number of steps that you wish to simulate the model. One
of two methods for deciding when to stop simulating.}

\item{method}{method for determining when to stop simulating. Can be set to
"Euler". One of two methods for deciding when to stop simulating.}
}
\description{
This function takes the model generated by the buildModel function, and uses that
as a basis to simulate the model
}
