% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{simulateNetwork}
\alias{simulateNetwork}
\title{A function to simulate a network}
\usage{
simulateNetwork(file, delay = 2, tmax = NA)
}
\arguments{
\item{file}{a file of the type produced by the buildModel function. Contains
the starting values for nodes, defines the genotypes of the
model, and defines the nextStep function for simulating the
change at each timestep.}

\item{delay}{the amount of delay to apply for delayed action of nodes. The
default is set to 2.}

\item{tmax}{the maximum number of steps that you want to simulate for. If
set to NA (the default), will simulate until stability is reached.}
}
\description{
Takes the output file of the buildNetwork function and uses it to run a
network simulation.
}
